// resources/js/types/pagination.d.ts
export interface Pagination {
current_page: number;
first_page_url?: string;
from?: number;
last_page: number;
last_page_url?: string;
links: Array<{ url: string | null; label: string; active: boolean }>;
next_page_url?: string;
path: string;
per_page: number;
prev_page_url?: string;
to?: number;
total: number;
}